refactor: SearchController 통합
